Control flow updates.

+++ deployment/www/stop_docker_stacks.sh
@@ -48,55 +48,51 @@ for DOMAIN_NAME in ${DOMAIN_LIST//,/ }; do
     done
 done
 
-# remove the nginx stack
-if docker stack list --format "{{.Name}}" | grep -q reverse-proxy; then
-    sleep 2
+if [ "$RESTART_FRONT_END" = true ]; then
+    # remove the nginx stack
+    if docker stack list --format "{{.Name}}" | grep -q reverse-proxy; then
+        sleep 2
 
-    docker stack rm reverse-proxy
+        docker stack rm reverse-proxy
 
-    # wait for all docker containers to stop.
-    # TODO see if there's a way to check for this.
-    sleep 15
-    
-fi
-
-# 
-if [ "$STOP_SERVICES" = true ]; then
-    echo "STOPPING as indicated by the --stop flag."
-    exit 0
-fi
-
-# generate the certs and grab a backup
-if [ "$RUN_CERT_RENEWAL" = true ]; then
-    ./generate_certs.sh
-fi
-
-# let's backup all our letsencrypt certs
-export APP="letsencrypt"
-for DOMAIN_NAME in ${DOMAIN_LIST//,/ }; do
-    export DOMAIN_NAME="$DOMAIN_NAME"
-    export SITE_PATH="$SITES_PATH/$DOMAIN_NAME"
-
-    # source the site path so we know what features it has.
-    source "$RESPOSITORY_PATH/reset_env.sh"
-    source "$SITE_PATH/site_definition"
-    source "$RESPOSITORY_PATH/domain_env.sh"
-
-    # these variable are used by both backup/restore scripts.
-    export REMOTE_BACKUP_PATH="$REMOTE_HOME/backups/www/$APP/$DOMAIN_IDENTIFIER"
-    export REMOTE_SOURCE_BACKUP_PATH="$REMOTE_HOME/$APP/$DOMAIN_NAME"
-
-    # ensure our local backup path exists so we can pull down the duplicity archive to the management machine.
-    export LOCAL_BACKUP_PATH="$SITE_PATH/backups/www/$APP"
-    mkdir -p "$LOCAL_BACKUP_PATH"
-
-    if [ "$RESTORE_WWW" = true ]; then
-        sleep 5
-        echo "STARTING restore_path.sh for letsencrypt."
-        ./restore_path.sh
-        #ssh "$PRIMARY_WWW_FQDN" sudo chown ubuntu:ubuntu "$REMOTE_HOME/$APP"
-    elif [ "$BACKUP_APPS"  = true ]; then
-        # if we're not restoring, then we may or may not back up.
-        ./backup_path.sh
+        # wait for all docker containers to stop.
+        # TODO see if there's a way to check for this.
+        sleep 15
+        
     fi
-done
+
+    # generate the certs and grab a backup
+    if [ "$RUN_CERT_RENEWAL" = true ]; then
+        ./generate_certs.sh
+    fi
+
+    # let's backup all our letsencrypt certs
+    export APP="letsencrypt"
+    for DOMAIN_NAME in ${DOMAIN_LIST//,/ }; do
+        export DOMAIN_NAME="$DOMAIN_NAME"
+        export SITE_PATH="$SITES_PATH/$DOMAIN_NAME"
+
+        # source the site path so we know what features it has.
+        source "$RESPOSITORY_PATH/reset_env.sh"
+        source "$SITE_PATH/site_definition"
+        source "$RESPOSITORY_PATH/domain_env.sh"
+
+        # these variable are used by both backup/restore scripts.
+        export REMOTE_BACKUP_PATH="$REMOTE_HOME/backups/www/$APP/$DOMAIN_IDENTIFIER"
+        export REMOTE_SOURCE_BACKUP_PATH="$REMOTE_HOME/$APP/$DOMAIN_NAME"
+
+        # ensure our local backup path exists so we can pull down the duplicity archive to the management machine.
+        export LOCAL_BACKUP_PATH="$SITE_PATH/backups/www/$APP"
+        mkdir -p "$LOCAL_BACKUP_PATH"
+
+        if [ "$RESTORE_WWW" = true ]; then
+            sleep 5
+            echo "STARTING restore_path.sh for letsencrypt."
+            ./restore_path.sh
+            #ssh "$PRIMARY_WWW_FQDN" sudo chown ubuntu:ubuntu "$REMOTE_HOME/$APP"
+        elif [ "$BACKUP_APPS"  = true ]; then
+            # if we're not restoring, then we may or may not back up.
+            ./backup_path.sh
+        fi
+    done
+fi
